move add gaol functionality in the gaol and progress page
